+++ CHANGELOG.md
@@ -11,6 +11,10 @@
  - Add `Packet<Dead>::dead_with_precision` to enable creating a pcap with nanosecond precision.
  - Add `flags` field to `Device`.
 
+### Removed
+
+- `BpfProgram` no longer have `Clone` implementation see [#261](https://github.com/rust-pcap/pcap/issues/261)
+
 ### Changed
 
 - MSRV is now `1.41.0`